+++ asset-transfer-events/application-gateway-go/go.mod
@@ -4,7 +4,7 @@ go 1.16
 
 require (
 	github.com/hyperledger/fabric-gateway v1.0.0
-	github.com/hyperledger/fabric-protos-go v0.0.0-20220125190318-19041b215616
+	github.com/hyperledger/fabric-protos-go v0.0.0-20220125190318-19041b215616 // indirect
 	github.com/miekg/pkcs11 v1.1.1 // indirect
 	golang.org/x/net v0.0.0-20220121210141-e204ce36a2ba // indirect
 	golang.org/x/sys v0.0.0-20220114195835-da31bd327af9 // indirect
